Switch to CSS-first styling
This is a big, breaking change. We remove the internal `theme` and
`style` concepts in favour of a CSS based styling approach.

This gives the user more control of the coloring and avoids having
two conflicting ways of making the same change.

It does remove some niche cases, which can be reintroduced by
popular request.